Rearrange TrsDataSync and DqtReporting bits
