+++ TeachingRecordSystem/src/TeachingRecordSystem.Core/Services/DqtReporting/DqtReportingOptions.cs
@@ -1,8 +1,8 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace TeachingRecordSystem.Core.Dqt.Services.TrsDataSync;
+namespace TeachingRecordSystem.Core.Services.DqtReporting;
 
-public class TrsDataSyncServiceOptions
+public class DqtReportingOptions
 {
     [Required]
     public required int PollIntervalSeconds { get; set; }
@@ -14,10 +14,10 @@ public class TrsDataSyncServiceOptions
     public required string CrmConnectionString { get; set; }
 
     [Required]
-    public required bool ProcessAllEntityTypesConcurrently { get; set; }
+    public required string ReportingDbConnectionString { get; set; }
 
     [Required]
-    public required bool IgnoreInvalidData { get; set; }
+    public required bool ProcessAllEntityTypesConcurrently { get; set; }
 
     [Required]
     public required bool RunService { get; set; }
