+++ TeachingRecordSystem/tests/TeachingRecordSystem.Core.Dqt.CrmIntegrationTests/Services/DqtReporting/DqtReportingFixture.cs
@@ -2,7 +2,7 @@ using Microsoft.ApplicationInsights;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Extensions.Options;
 using Microsoft.Xrm.Sdk;
-using TeachingRecordSystem.Core.Dqt.Services.DqtReporting;
+using TeachingRecordSystem.Core.Services.DqtReporting;
 
 namespace TeachingRecordSystem.Core.Dqt.CrmIntegrationTests.Services.DqtReporting;
