+++ TeachingRecordSystem/tests/TeachingRecordSystem.Core.Dqt.CrmIntegrationTests/TeachingRecordSystem.Core.Dqt.CrmIntegrationTests.csproj
@@ -23,8 +23,6 @@
     <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" />
     <PackageReference Include="Moq" />
-    <PackageReference Include="System.Linq.Async" />
-    <PackageReference Include="System.Reactive" />
     <PackageReference Include="xunit" />
     <PackageReference Include="Xunit.DependencyInjection" />
     <PackageReference Include="xunit.runner.visualstudio">
