+++ TeachingRecordSystem/tests/TeachingRecordSystem.Core.Tests/Services/TrsDataSync/TrsDataSyncHelperTests.cs
@@ -1,23 +1,27 @@
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Xrm.Sdk;
+using TeachingRecordSystem.Core.Dqt.Models;
+using TeachingRecordSystem.Core.Services.TrsDataSync;
 
-namespace TeachingRecordSystem.Core.Dqt.CrmIntegrationTests.Services.TrsDataSync;
+namespace TeachingRecordSystem.Core.Tests.Services.TrsDataSync;
 
-public class TrsDataSyncServiceTests : IClassFixture<TrsDataSyncServiceFixture>
+public class TrsDataSyncHelperTests
 {
-    private readonly TrsDataSyncServiceFixture _fixture;
+    private readonly DbFixture _dbFixture;
     private readonly FakeTrnGenerator _trnGenerator;
 
-    public TrsDataSyncServiceTests(TrsDataSyncServiceFixture fixture, FakeTrnGenerator trnGenerator)
+    public TrsDataSyncHelperTests(DbFixture dbFixture, FakeTrnGenerator trnGenerator)
     {
-        _fixture = fixture;
+        _dbFixture = dbFixture;
         _trnGenerator = trnGenerator;
     }
 
     [Fact]
-    public async Task ProcessChangesForEntityType_WritesNewRecordToDatabase()
+    public async Task SyncContact_NewRecord_WritesNewRowToDb()
     {
         // Arrange
+        var clock = new TestableClock();
+        var helper = new TrsDataSyncHelper(_dbFixture.GetDbContextFactory(), clock);
+
         var contactId = Guid.NewGuid();
         var trn = _trnGenerator.GenerateTrn();
         var firstName = Faker.Name.First();
@@ -26,8 +30,8 @@ public class TrsDataSyncServiceTests : IClassFixture<TrsDataSyncServiceFixture>
         var dateOfBirth = Faker.Identification.DateOfBirth();
         var email = Faker.Internet.Email();
         var nino = Faker.Identification.UkNationalInsuranceNumber();
-        var created = _fixture.Clock.UtcNow;
-        var modified = _fixture.Clock.Advance();
+        var created = clock.UtcNow;
+        var modified = clock.Advance();
 
         var contact = new Contact()
         {
@@ -47,13 +51,11 @@ public class TrsDataSyncServiceTests : IClassFixture<TrsDataSyncServiceFixture>
             ModifiedOn = modified
         };
 
-        var newItem = new NewOrUpdatedItem(ChangeType.NewOrUpdated, contact);
-
         // Act
-        await _fixture.PublishChangedItemAndConsume(newItem);
+        await helper.SyncContact(contact, ignoreInvalid: false);
 
         // Assert
-        await _fixture.DbFixture.WithDbContext(async dbContext =>
+        await _dbFixture.WithDbContext(async dbContext =>
         {
             var person = await dbContext.Persons.SingleOrDefaultAsync(p => p.DqtContactId == contact.Id);
             Assert.NotNull(person);
@@ -69,15 +71,18 @@ public class TrsDataSyncServiceTests : IClassFixture<TrsDataSyncServiceFixture>
             Assert.Equal((int)ContactState.Active, person.DqtState);
             Assert.Equal(created, person.DqtCreatedOn);
             Assert.Equal(modified, person.DqtModifiedOn);
-            Assert.Equal(_fixture.Clock.UtcNow, person.DqtFirstSync);
-            Assert.Equal(_fixture.Clock.UtcNow, person.DqtLastSync);
+            Assert.Equal(clock.UtcNow, person.DqtFirstSync);
+            Assert.Equal(clock.UtcNow, person.DqtLastSync);
         });
     }
 
     [Fact]
-    public async Task ProcessChangesForEntityType_WritesUpdatedRecordToDatabase()
+    public async Task SyncContact_ExistingRecord_UpdatesExistingRowInDb()
     {
         // Arrange
+        var clock = new TestableClock();
+        var helper = new TrsDataSyncHelper(_dbFixture.GetDbContextFactory(), clock);
+
         var contactId = Guid.NewGuid();
         var trn = _trnGenerator.GenerateTrn();
 
@@ -87,11 +92,11 @@ public class TrsDataSyncServiceTests : IClassFixture<TrsDataSyncServiceFixture>
         var initialDateOfBirth = Faker.Identification.DateOfBirth();
         var initialEmail = Faker.Internet.Email();
         var initialNino = Faker.Identification.UkNationalInsuranceNumber();
-        var initialSyncTime = _fixture.Clock.UtcNow;
-        var created = _fixture.Clock.UtcNow;
-        var initialModified = _fixture.Clock.UtcNow;
+        var initialSyncTime = clock.UtcNow;
+        var created = clock.UtcNow;
+        var initialModified = clock.UtcNow;
 
-        await _fixture.DbFixture.WithDbContext(async dbContext =>
+        await _dbFixture.WithDbContext(async dbContext =>
         {
             dbContext.Persons.Add(new()
             {
@@ -114,7 +119,7 @@ public class TrsDataSyncServiceTests : IClassFixture<TrsDataSyncServiceFixture>
             await dbContext.SaveChangesAsync();
         });
 
-        _fixture.Clock.Advance();
+        clock.Advance();
 
         var newFirstName = Faker.Name.First();
         var newMiddleName = Faker.Name.Middle();
@@ -122,7 +127,7 @@ public class TrsDataSyncServiceTests : IClassFixture<TrsDataSyncServiceFixture>
         var newDateOfBirth = Faker.Identification.DateOfBirth();
         var newEmail = Faker.Internet.Email();
         var newNino = Faker.Identification.UkNationalInsuranceNumber();
-        var newModified = _fixture.Clock.UtcNow;
+        var newModified = clock.UtcNow;
 
         var contact = new Contact()
         {
@@ -142,13 +147,11 @@ public class TrsDataSyncServiceTests : IClassFixture<TrsDataSyncServiceFixture>
             dfeta_StatedLastName = null
         };
 
-        var updatedItem = new NewOrUpdatedItem(ChangeType.NewOrUpdated, contact);
-
         // Act
-        await _fixture.PublishChangedItemAndConsume(updatedItem);
+        await helper.SyncContact(contact, ignoreInvalid: false);
 
         // Assert
-        await _fixture.DbFixture.WithDbContext(async dbContext =>
+        await _dbFixture.WithDbContext(async dbContext =>
         {
             var person = await dbContext.Persons.SingleOrDefaultAsync(p => p.DqtContactId == contact.Id);
             Assert.NotNull(person);
@@ -162,14 +165,17 @@ public class TrsDataSyncServiceTests : IClassFixture<TrsDataSyncServiceFixture>
             Assert.Equal(created, person.DqtCreatedOn);
             Assert.Equal(newModified, person.DqtModifiedOn);
             Assert.Equal(initialSyncTime, person.DqtFirstSync);
-            Assert.Equal(_fixture.Clock.UtcNow, person.DqtLastSync);
+            Assert.Equal(clock.UtcNow, person.DqtLastSync);
         });
     }
 
     [Fact]
-    public async Task ProcessChangesForEntityType_DeletesRemovedRecordFromDatabase()
+    public async Task DeleteEntities_RemovesRowFromDb()
     {
         // Arrange
+        var clock = new TestableClock();
+        var helper = new TrsDataSyncHelper(_dbFixture.GetDbContextFactory(), clock);
+
         var contactId = Guid.NewGuid();
         var trn = _trnGenerator.GenerateTrn();
 
@@ -179,9 +185,9 @@ public class TrsDataSyncServiceTests : IClassFixture<TrsDataSyncServiceFixture>
         var dateOfBirth = Faker.Identification.DateOfBirth();
         var email = Faker.Internet.Email();
         var nino = Faker.Identification.UkNationalInsuranceNumber();
-        var syncTime = _fixture.Clock.UtcNow;
+        var syncTime = clock.UtcNow;
 
-        await _fixture.DbFixture.WithDbContext(async dbContext =>
+        await _dbFixture.WithDbContext(async dbContext =>
         {
             dbContext.Persons.Add(new()
             {
@@ -202,13 +208,11 @@ public class TrsDataSyncServiceTests : IClassFixture<TrsDataSyncServiceFixture>
             await dbContext.SaveChangesAsync();
         });
 
-        var removedItem = new RemovedOrDeletedItem(ChangeType.RemoveOrDeleted, new EntityReference(Contact.EntityLogicalName, contactId));
-
         // Act
-        await _fixture.PublishChangedItemAndConsume(removedItem);
+        await helper.DeleteEntities(Contact.EntityLogicalName, new[] { contactId });
 
         // Assert
-        await _fixture.DbFixture.WithDbContext(async dbContext =>
+        await _dbFixture.WithDbContext(async dbContext =>
         {
             var person = await dbContext.Persons.SingleOrDefaultAsync(p => p.DqtContactId == contactId);
             Assert.Null(person);
