+++ TeachingRecordSystem/tests/TeachingRecordSystem.Core.Tests/Startup.cs
@@ -1,6 +1,7 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using TeachingRecordSystem.Core.Dqt;
 
 namespace TeachingRecordSystem.Core.Tests;
 
@@ -17,5 +18,7 @@ public class Startup
 
                 services.AddSingleton<DbFixture>();
                 services.AddSingleton<FakeTrnGenerator>();
+                services.AddCrmQueries();
+                services.AddFakeXrm();
             });
 }
