+++ TeachingRecordSystem/tests/TeachingRecordSystem.TestCommon/TeachingRecordSystem.TestCommon.csproj
@@ -15,6 +15,8 @@
     <PackageReference Include="FakeXrmEasy.v9" />
     <PackageReference Include="Moq" />
     <PackageReference Include="Respawn" />
+    <PackageReference Include="System.Linq.Async" />
+    <PackageReference Include="System.Reactive" />
     <PackageReference Include="xunit.assert" />
   </ItemGroup>
